supabase/functions/vectorize-image/requirements.txt
supabase-functions
fastapi
pydantic
opencv-python
numpy
svgwrite
Pillow